chore: structure the core functionality differently, this might change in the future
